class-32/code_review/blogapi-permissions/posts/serializers.py
from rest_framework import serializers
from .models import Post


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        fields = (
            "id",
            "authors",
            "title",
            "body",
        )
        model = Post
